+++ src/utils/utils.ts
@@ -25,7 +25,6 @@ export const setFeaturedItemArray = () => {
  * Returns a random image ID from the list of featured items.
  */
 export const generateRandomImageID = () => {
-  console.log("generating random imageID");
   if (featuredItemArray.length === 0) {
     setFeaturedItemArray();
   }
@@ -43,7 +42,6 @@ export const featuredImageID = (imageID = "") => {
   if (featuredItemArray.length === 0) {
     setFeaturedItemArray();
   }
-
   if (imageID !== "") {
     return featuredItemArray.includes(imageID)
       ? imageID
@@ -79,6 +77,10 @@ export const imageURL = (
 export const getNumDigitizedItems = async () => {
   const apiUrl = `${process.env.API_URL}/api/v2/items/total`;
   // const res = await apiCall(apiUrl);
+  console.log(`getNumDigitizedItems: About to fetch ${apiUrl}`);
+  console.log(`getNumDigitizedItems calls apiCall function internally`);
+  // const res = await apiCall(apiUrl);
+
   const fallbackCount = 863848;
   // const totalItems = res.count.$ || fallbackCount;
   const totalItems = fallbackCount;
@@ -92,7 +94,10 @@ export const getNumDigitizedItems = async () => {
 
 export const getNumItems = async (uuid: string) => {
   const apiUrl = `${process.env.API_URL}/api/v2/collections/${uuid}/items`;
+  console.log(`getNumItems: About to fetch ${apiUrl}`);
+  console.log(`getNumItems calls apiCall function internally`);
   // const res = await apiCall(apiUrl);
+  // return res.numItems || 0;
   return 0; //res.numItems || 0;
 };
 
@@ -104,7 +109,10 @@ export const getNumItems = async (uuid: string) => {
 
 export const getAPIUri = async (identifierType: string, identifier: string) => {
   const apiUrl = `${process.env.API_URL}/api/v2/items/${identifierType}/${identifier}`;
-  return apiCall(apiUrl);
+  console.log(`getAPIUri: About to fetch ${apiUrl}`);
+  console.log(`getAPIUri calls apiCall function internally`);
+  const apiCallValue = apiCall(apiUrl);
+  return apiCallValue;
 };
 
 /**
@@ -115,6 +123,7 @@ export const getAPIUri = async (identifierType: string, identifier: string) => {
 export const apiCall = async (apiUrl: string) => {
   const apiKey = process.env.AUTH_TOKEN;
   try {
+    const startTime = new Date().getTime();
     const response = await fetch(apiUrl, {
       headers: {
         Authorization: `Token token=${apiKey}`,
@@ -123,6 +132,8 @@ export const apiCall = async (apiUrl: string) => {
 
     if (response.status === 200) {
       const data = await response.json();
+      console.log(`apiCall: called ${apiUrl}`);
+      console.log(`Response time: ${new Date().getTime() - startTime}`);
       return data.nyplAPI.response;
     } else {
       return undefined;
@@ -139,6 +150,11 @@ export const getItemDataFromImageID = async (imageID: string) => {
   //   uuid: apiUri.uuid,
   //   title: getTitleFromRepoAPIResponseData(data) || "",
   // };
+  console.log(
+    `getItemDataFromImageID: About call getAPIUri and apiCall for image id: ${imageID}`
+  );
+  // const apiUri = await getAPIUri("local_image_id", imageID);
+  // const data = await apiCall(apiUri.apiUri);
   return {
     uuid: "6265a5c0-c5ef-012f-687c-58d385a7bc34", //apiUri.uuid,
     title: "Greetings from Staten Island",
@@ -167,4 +183,23 @@ function getTitleFromRepoAPIResponseData(data) {
     title = titleInfo.title.$;
   }
   return title;
+}
+
+export function getCustomTimestamp() {
+  // Return the formatted number
+  // timestamp: date
+  const currentDate = new Date();
+  const currentDayOfMonth = currentDate.getDate();
+  const currentMonth = currentDate.getMonth(); // Be careful! January is 0, not 1
+  const currentYear = currentDate.getFullYear();
+  const dateString =
+    currentYear + "-" + (currentMonth + 1) + "-" + currentDayOfMonth;
+  // timestamp: time
+  const hh = currentDate.getUTCHours();
+  const mm = currentDate.getUTCMinutes();
+  const ss = currentDate.getSeconds();
+  const timeString = hh + ":" + mm + ":" + ss;
+  const timestamp = dateString + " " + timeString;
+
+  return timestamp;
 }