+++ src/pages/api/featuredItem.ts
@@ -39,6 +39,9 @@ const featuredItemDataHandler = async (
       }`,
     };
 
+    // 24 hour cache
+    response.setHeader("Cache-Control", "s-maxage=86400");
+
     return response.status(200).json({
       featuredItem: featuredItemObject,
       numberOfDigitizedItems: numDigitizedItems,