+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated favicon.ico
 - Added feedback button and logic for form to submit to google spreadsheet (DR-2595)
 
+### Added
+
+- Added AA Data Layer on the home page (DR-2700)
+
 ## [0.1.0] 2024-01-08
 
 ### Created