+++ CHANGELOG.md
@@ -10,12 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 
 - Updated favicon.ico
-- Added feedback button and logic for form to submit to google spreadsheet (DR-2595)
+- Skeleton load the campaign hero (DR-2787)
+- Skeleton load the swim lanes (DR-2788)
 
 ### Added
 
 - Added Optinmonster embed code (DR-2744)
 - Added AA Data Layer on the home page (DR-2700)
+- Added feedback button and logic for form to submit to google spreadsheet (DR-2595)
 
 ## [0.1.0] 2024-01-08