+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Optinmonster embed code (DR-2744)
 - Added AA Data Layer on the home page (DR-2700)
 - Added feedback button and logic for form to submit to google spreadsheet (DR-2595)
+- Added newrelic agent (DR-2772)
+- Added canonical link (DR-2784)
 
 ## [0.1.0] 2024-01-08