+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.1] 2024-03-13
+
+### Updated
+
+- Updated favicon.ico
+- Skeleton load the campaign hero (DR-2787)
+- Skeleton load the swim lanes (DR-2788)
+- Use Repo API v2/items/featured endpoint when loading images for hero component (DR-2801)
+
+### Added
+
+- Added Optinmonster embed code (DR-2744)
+- Added AA Data Layer on the home page (DR-2700)
+- Added feedback button and logic for form to submit to google spreadsheet (DR-2595)
+- Added newrelic agent (DR-2772)
+- Added canonical link (DR-2784)
+- Added caching to API routes for `/api/featuredItem ` and `/api/homepage` as well as from the client side (DR-2785)
+
 ## [0.1.0] 2024-01-08
 
 ### Created